Fix: Discrepancy in current turn player
When the current player is the last player in the trick
(i.e. the next player is the lead player)
it means this current player will end the trick.
There is no need to advance the turn since he
true next player will be determined later by
computing the trick winner.
This test keeps the ui more consistent
if the player who wins the trick is the last player in the trick.