Fix: premature transition to GameOverPhase
Declaring a winner based on highest/lowest trump in hand gives the "losing" team an unfair insight into the range of trump their opponents might have.

For example: suppose your opponents have 13 points. Currently, when the round
begins the computer checks your opponents have the necessary sure points to
win the game and if they do then the game is shown to be over. So if the game
is not over in this scenario it means that you can be sure that your opponents
do not have high and hence you have high (assuming your team is holding any
trump).