Fix round summary calculation (#7)
`compare` Function Misbehavior

After merging the Rescript 11 migration. Players noticed that the awards for High and Low Trump were incorrect. I suspect that the JS code that Rescript 11 generated for Card.Rank.t was different enough from the Rescript 10 output that it caused the compare function to misbehave.

Now instead of comparing the Rank.t values directly, I'm first converting them to an integer representation of the value of the Rank and then comparing them.